pyproject.toml
[project]
name = "otocyon"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "duckdb>=1.5.1",
    "polars>=1.39.3",
    "ruamel-yaml>=0.19.1",
    "numpy",
    "pyarrow>=23.0.1",
    "msgspec>=0.19.0",
    "structlog>=25.0.0",
    "cyclopts>=3.0.0",
    "pandera[polars]>=0.22.0",
]

[project.scripts]
otocyon = "otocyon.cli:app"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "ruff>=0.15.8",
    "beartype>=0.20.0",
    "mypy>=1.15.0",
    "hypothesis>=6.130.0",
    "pytest-benchmark>=5.1.0",
]
docs = [
    "mkdocstrings[python]>=1.0.3",
    "zensical>=0.0.31",
]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
filterwarnings = [
    # pandera's polars backend uses `col in LazyFrame` internally — not our code
    "ignore::polars.exceptions.PerformanceWarning:pandera",
]

[tool.pyright]
venvPath = "."
venv = ".venv"

[tool.mypy]
python_version = "3.13"
strict = true
ignore_missing_imports = true
plugins = []
